// src/payload-types.ts
export interface Media {
  id: number
  alt: string
  folder?: ('posts' | 'avatars' | 'others') | null
  updatedAt: string
  createdAt: string
  url?: string | null
  thumbnailURL?: string | null
  filename?: string | null
  mimeType?: string | null
  filesize?: number | null
  width?: number | null
  height?: number | null
  focalX?: number | null
  focalY?: number | null
}
// ...
export interface Brand {
  id: number
  name: string
  slug: string
  logo?: (number | null) | Media
  description?: string | null
  asin_count?: number | null
  topics_count?: number | null
  click_share_sum?: number | null
  conversion_share_sum?: number | null
  weighted_score_sum?: number | null
  updatedAt: string
  createdAt: string
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "products".
 */
export interface Product {
  id: number
  asin: string
  brand?: (number | null) | Brand
  active?: boolean | null
  image?: string | null
  paapi5?:
    | {
        [k: string]: unknown
      }
    | unknown[]
    | string
    | number
    | boolean
    | null
  status: 'PENDING' | 'PROCESSING' | 'COMPLETED' | 'ERROR' | 'NOT_FOUND'
  message?: string | null
  completed_at?: string | null
  click_share_sum?: number | null
  conversion_share_sum?: number | null
  weighted_score_sum?: number | null
  updatedAt: string
  createdAt: string
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "product-scrapers".
 */
export interface ProductScraper {
  id: number
  asin: string
  active?: boolean | null
  scraperMetadata?:
    | {
        [k: string]: unknown
      }
    | unknown[]
    | string
    | number
    | boolean
    | null
  status: 'PENDING' | 'PROCESSING' | 'COMPLETED' | 'ERROR' | 'NOT_FOUND'
  message?: string | null
  completed_at?: string | null
  updatedAt: string
  createdAt: string
}
// ...
export interface AbaSearchTerm {
  id: number
  keywords: string
  asin: string
  /**
   * Node IDs
   */
  node_ids?:
    | {
        [k: string]: unknown
      }
    | unknown[]
    | string
    | number
    | boolean
    | null
  /**
   * Occurrences
   */
  occurrences?: number | null
  /**
   * Search Rank
   */
  search_rank?: number | null
  /**
   * Conversion Share
   */
  conversion_share?: number | null
  /**
   * Click Share
   */
  click_share?: number | null
  /**
   * Weighted Score
   */
  weighted_score?: number | null
  start_date?: string | null
  end_date?: string | null
  updatedAt: string
  createdAt: string
}